Split Nuklear into its rendering part and add DebugView system
