+++ src/Application/DebugView.h
@@ -1,0 +1,9 @@
+#pragma once
+
+class Nuklear;
+
+class DebugView {
+ public:
+    virtual ~DebugView() = default;
+    virtual void update(Nuklear &nuklear) = 0;
+};
