+++ src/Application/GameStarter.h
@@ -18,6 +18,7 @@ class Engine;
 class Game;
 class GameLuaBindings;
 class ScriptingSystem;
+class DebugViewSystem;
 class EngineController;
 
 class GameStarter {
@@ -51,6 +52,7 @@ class GameStarter {
     std::unique_ptr<PlatformApplication> _application;
     std::unique_ptr<Renderer> _renderer;
     std::unique_ptr<ScriptingSystem> _scriptingSystem;
+    std::unique_ptr<DebugViewSystem> _debugViewSystem;
     std::unique_ptr<Nuklear> _nuklear;
     std::unique_ptr<Engine> _engine;
     std::shared_ptr<Game> _game;
