+++ src/Application/RenderStatsDebugView.h
@@ -1,0 +1,15 @@
+#pragma once
+
+#include "DebugView.h"
+
+class Renderer;
+
+class RenderStatsDebugView : public DebugView {
+ public:
+    explicit RenderStatsDebugView(Renderer &renderer);
+
+    void update(Nuklear &nuklear) override;
+
+ private:
+    Renderer& _renderer;
+};
