+++ src/Engine/Engine.h
@@ -92,6 +92,7 @@ class Engine {
     void Draw();
     void drawWorld();
     void drawHUD();
+    void drawDebugView();
     void DrawGUI();
     void ResetCursor_Palettes_LODs_Level_Audio_SFT_Windows();
     void SecondaryInitialization();
