+++ src/Engine/Graphics/CMakeLists.txt
@@ -19,8 +19,6 @@ set(ENGINE_GRAPHICS_SOURCES
         LightsStack.cpp
         LocationFunctions.cpp
         Nuklear.cpp
-        NuklearEventHandler.cpp
-        NuklearUtils.cpp
         Outdoor.cpp
         Overlays.cpp
         PaletteManager.cpp
@@ -53,8 +51,6 @@ set(ENGINE_GRAPHICS_HEADERS
         LocationInfo.h
         LocationTime.h
         Nuklear.h
-        NuklearEventHandler.h
-        NuklearUtils.h
         Outdoor.h
         Overlays.h
         PaletteManager.h
