+++ src/Engine/Graphics/Renderer/CMakeLists.txt
@@ -2,6 +2,7 @@ cmake_minimum_required(VERSION 3.27 FATAL_ERROR)
 
 set(ENGINE_GRAPHICS_RENDERER_SOURCES
         BaseRenderer.cpp
+        NuklearDebugViewRenderer.cpp
         NullRenderer.cpp
         OpenGLRenderer.cpp
         OpenGLShader.cpp
@@ -11,7 +12,8 @@ set(ENGINE_GRAPHICS_RENDERER_SOURCES
         )
 
 set(ENGINE_GRAPHICS_RENDERER_HEADERS
-        BaseRenderer.h
+        BaseRenderer.h        
+        IDebugViewRenderer.h
         NullRenderer.h
         OpenGLRenderer.h
         OpenGLShader.h
