+++ src/Engine/Graphics/Renderer/OpenGLRenderer.h
@@ -31,15 +31,6 @@ class OpenGLRenderer : public BaseRenderer {
 
     virtual bool Initialize() override;
 
-    virtual bool NuklearInitialize(struct nk_tex_font *tfont) override;
-    virtual bool NuklearCreateDevice() override;
-    virtual bool NuklearRender(/*enum nk_anti_aliasing*/ int AA, int max_vertex_buffer, int max_element_buffer) override;
-    virtual void NuklearRelease() override;
-    virtual struct nk_tex_font *NuklearFontLoad(const char *font_path, size_t font_size) override;
-    virtual void NuklearFontFree(struct nk_tex_font *tfont) override;
-    virtual struct nk_image NuklearImageLoad(GraphicsImage *img) override;
-    virtual void NuklearImageFree(GraphicsImage *img) override;
-
     virtual RgbaImage ReadScreenPixels() override;
     virtual void ClearTarget(Color uColor) override;
     virtual void Present() override;
@@ -179,7 +170,6 @@ class OpenGLRenderer : public BaseRenderer {
     OpenGLShader billbshader;
     OpenGLShader decalshader;
     OpenGLShader forcepershader;
-    OpenGLShader nuklearshader;
 
     // terrain shader
     GLuint terrainVBO{}, terrainVAO{};
@@ -232,8 +222,6 @@ class OpenGLRenderer : public BaseRenderer {
     int fogend{};
 
     float gamma{};
-
-    std::unique_ptr<nk_state> nk;
 };
