+++ src/GUI/UI/UIGame.cpp
@@ -1829,11 +1829,7 @@ GUIWindow_DebugMenu::GUIWindow_DebugMenu()
     GUIButton *pBtn_DebugSpecialItem = CreateButton({354, 275}, {width, height}, 1, 0, UIMSG_DebugSpecialItem, 0, Io::InputAction::Invalid, "DEBUG GENERATE RANDOM SPECIAL ITEM");
 
     GUIButton *pBtn_DebugReloadShaders = CreateButton({13, 302}, {width, height}, 1, 0, UIMSG_DebugReloadShader, 0, Io::InputAction::ReloadShaders, "DEBUG RELOAD SHADERS");
-    if (nuklear->isInitialized(WindowType::WINDOW_DebugMenu)) {
-        GUIButton* pBtn_DebugUnused1 = CreateButton({ 127, 302 }, { width, height }, 1, 0, UIMSG_DebugConsole, 0, Io::InputAction::Invalid, "DEBUG TOGGLE CONSOLE");
-    } else {
-        GUIButton* pBtn_DebugUnused1 = CreateButton({ 127, 302 }, { width, height }, 1, 0, UIMSG_DebugUnused, 0, Io::InputAction::Invalid, "DEBUG unused1");
-    }
+    GUIButton* pBtn_DebugUnused1 = CreateButton({ 127, 302 }, { width, height }, 1, 0, UIMSG_DebugConsole, 0, Io::InputAction::Invalid, "DEBUG TOGGLE CONSOLE");
     GUIButton *pBtn_DebugUnused2 = CreateButton({241, 302}, {width, height}, 1, 0, UIMSG_DebugUnused, 0, Io::InputAction::Invalid, "DEBUG unused2");
     GUIButton *pBtn_DebugUnused3 = CreateButton({354, 302}, {width, height}, 1, 0, UIMSG_DebugUnused, 0, Io::InputAction::Invalid, "DEBUG unused3");
 
@@ -1892,11 +1888,7 @@ void GUIWindow_DebugMenu::Update() {
     buttonbox(354, 275, "Special Item", 2);
 
     buttonbox(13, 302, "HOT Shaders", 2);
-    if (nuklear->isInitialized(WindowType::WINDOW_DebugMenu)) {
-        buttonbox(127, 302, "Console", engine->config->debug.ShowConsole.value());
-    } else {
-        buttonbox(127, 302, "Unused1", 2);
-    }
+    buttonbox(127, 302, "Console", engine->config->debug.ShowConsole.value());
     buttonbox(241, 302, "Unused2", 2);
     buttonbox(354, 302, "Unused3", 2);
